requirements.txt
numpy==1.19.5
pandas==1.1.5
seaborn==0.11.2
matplotlib==3.3.4
scikit-image==0.17.2
scikit-learn==0.24.2
Pillow==8.4.0
opencv-python==4.6.0.66
openpyxl==3.0.10
catalyst==20.12
imagesize==1.3.0
torch==1.10.1
torchvision==0.11.2
timm==0.5.4
excel==1.0.0
xlrd==1.2.0
albumentations==1.1.0
tqdm==4.64.0